fix potential type mismatch between format specifier and argument
Cast printf argument of type time_t to long long to match the %lld
format specifier on platforms where this might not be the case.

+++ lib/sigs.c
@@ -202,7 +202,7 @@ got_tag_write_signed_data(BUF *buf, struct got_tag_object *tag,
 	got_date_format_gmtoff(gmtoff, sizeof(gmtoff),
 	    got_object_tag_get_tagger_gmtoff(tag));
 	if (asprintf(&tagger, "%s %lld %s", got_object_tag_get_tagger(tag),
-	    got_object_tag_get_tagger_time(tag), gmtoff) == -1) {
+	    (long long)got_object_tag_get_tagger_time(tag), gmtoff) == -1) {
 		err = got_error_from_errno("asprintf");
 		goto done;
 	}
